split screen and stages classes, remove final modifier
